🎉  (tailwind) NICE-35 more prep for multi-site
