+++ packages/shared/package.json
@@ -31,7 +31,6 @@
   },
   "dependencies": {
     "@jeromefitz/design-system": "4.2.0",
-    "@stitches/react": "1.3.1-1",
     "@tanem/react-nprogress": "5.0.34",
     "fathom-client": "3.5.0",
     "lodash": "4.17.21",
