+++ packages/tailwind-config/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@jeromefitz/tailwind-config",
+  "version": "0.0.0",
+  "private": true,
+  "devDependencies": {
+    "@plaiceholder/tailwindcss": "2.5.0",
+    "@radix-ui/colors": "0.1.8",
+    "@tailwindcss/aspect-ratio": "0.4.2",
+    "@tailwindcss/forms": "0.5.3",
+    "@tailwindcss/typography": "0.5.9",
+    "autoprefixer": "10.4.14",
+    "cssnano": "6.0.0",
+    "postcss": "8.4.23",
+    "radix-colors-for-tailwind": "2.0.0",
+    "tailwindcss": "3.3.1",
+    "tailwindcss-animate": "1.0.5",
+    "tailwindcss-radix": "2.8.0"
+  }
+}
