+++ packages/tailwind-config/styles/globals.css
