+++ sites/jeromefitzgerald.com/package.json
@@ -41,7 +41,6 @@
     "@jeromefitz/spotify": "3.0.2",
     "@jeromefitz/utils": "2.2.2",
     "@mantine/hooks": "6.0.8",
-    "@radix-ui/colors": "0.1.8",
     "@radix-ui/primitive": "1.0.0",
     "@radix-ui/react-accessible-icon": "1.0.2",
     "@radix-ui/react-accordion": "1.1.1",
@@ -96,8 +95,6 @@
     "@radix-ui/react-use-rect": "1.0.0",
     "@radix-ui/react-use-size": "1.0.0",
     "@radix-ui/react-visually-hidden": "1.0.2",
-    "@stitches/react": "1.3.1-1",
-    "@tailwindcss/forms": "0.5.3",
     "@tanem/react-nprogress": "5.0.34",
     "@upstash/ratelimit": "0.4.2",
     "@upstash/redis": "1.20.4",
@@ -105,7 +102,6 @@
     "@vercel/og": "0.5.3",
     "clsx": "1.2.1",
     "cmdk": "0.2.0",
-    "cssnano": "6.0.0",
     "date-fns": "2.29.3",
     "date-fns-tz": "2.0.0",
     "fast-json-stable-stringify": "2.1.0",
@@ -143,10 +139,8 @@
     "@jeromefitz/lighthouse-config": "workspace:*",
     "@jeromefitz/next-config": "workspace:*",
     "@jeromefitz/shared": "workspace:*",
+    "@jeromefitz/tailwind-config": "workspace:*",
     "@jeromefitz/tsconfig": "1.1.3",
-    "@plaiceholder/tailwindcss": "2.5.0",
-    "@tailwindcss/aspect-ratio": "0.4.2",
-    "@tailwindcss/typography": "0.5.9",
     "@types/ms": "0.7.31",
     "@types/pluralize": "0.0.29",
     "@types/ramda": "0.29.0",
@@ -154,15 +148,9 @@
     "@types/react-dom": "18.0.11",
     "@types/title": "3.4.1",
     "@types/uuid": "9.0.1",
-    "autoprefixer": "10.4.14",
     "next-notion": "workspace:*",
     "next-unused": "0.0.6",
-    "postcss": "8.4.23",
-    "radix-colors-for-tailwind": "2.0.0",
-    "tailwind-merge": "1.12.0",
-    "tailwindcss": "3.3.1",
-    "tailwindcss-animate": "1.0.5",
-    "tailwindcss-radix": "2.8.0"
+    "tailwind-merge": "1.12.0"
   },
   "next-unused": {
     "alias": {
