+++ sites/jeromefitzgerald.com/postcss.config.js
@@ -1,7 +1,1 @@
-module.exports = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-    ...(process.env.NODE_ENV === 'production' ? { cssnano: {} } : {}),
-  },
-}
+module.exports = require('@jeromefitz/tailwind-config/postcss.config')
